safari-600.7-branch/LayoutTests/pointer-lock/mouse-event-api.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../http/tests/resources/js-test-pre.js"></script>
</head>
<body>
<script>
    var mouseEvent = document.createEvent("MouseEvent");
    shouldBeDefined("mouseEvent.movementX");
    shouldBeDefined("mouseEvent.movementY");
</script>
<script src="../http/tests/resources/js-test-post.js"></script>
</body>
</html>
